backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai==1.3.0
langchain-text-splitters==0.0.1
PyPDF2==3.0.1
PyMuPDF==1.23.8
llama-parse==0.1.4
python-multipart==0.0.6
Pillow==10.1.0 